requirements-dev.txt
pydantic==2.7.1
pytest==8.2.0
pytest-mock==3.14.0
ruff==0.4.4
black==24.4.2
yamllint==1.35.1
